Add a flip
